fix chat_id, make it BigInteger

+++ bot/db.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, Time, DateTime, create_engine
+from sqlalchemy import Column, Integer, String, Date, Time, DateTime, BigInteger, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 import os
@@ -11,7 +11,7 @@ Base = declarative_base()
 
 class Chat(Base):
     __tablename__ = 'chats'
-    chat_id = Column(Integer, primary_key=True)
+    chat_id = Column(BigInteger, primary_key=True)
     data_begin = Column(Date)
     month_budget = Column(Integer)
     balance = Column(Integer)
